# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Haley's AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  model: string;
  createdAt: string;
  messageCount: number;
}

export interface ConversationInput {
  title: string;
  model?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  model: string;
  createdAt: string;
  messages: Message[];
}

export interface ChatInput {
  message: string;
  conversationId: number;
  model?: string;
}

export interface ChatResponse {
  reply: string;
  conversationId: number;
  messageId: number;
}

export interface FileEntry {
  name: string;
  path: string;
  type: string;
  size: number;
}

export interface FileInput {
  path: string;
  content: string;
}

export interface FileContent {
  path: string;
  content: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface CommandInput {
  command: string;
  cwd?: string;
}

export interface CommandResult {
  stdout: string;
  stderr: string;
  exitCode: number;
}

export interface Stats {
  totalConversations: number;
  totalMessages: number;
  totalFiles: number;
  modelsUsed: string[];
}

export type ListFilesParams = {
path?: string;
};

export type ReadFileParams = {
path: string;
};

export type DeleteFileParams = {
path: string;
};