# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Haley's AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  reply: string;
  conversationId: number;
  messageId: number;
}