# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Haley's AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalConversations: number;
  totalMessages: number;
  totalFiles: number;
  modelsUsed: string[];
}